Split bashrc from profile

+++ install.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+make all
+
+PREFIX="${PREFIX:-/}"
+
 if [ ! -w "$PREFIX/" ]; then
 	printf "Can't access prefix directory: %s\n" "$PREFIX/" >&2
 	exit 1
@@ -10,9 +14,9 @@ if [ -z "$BINPREFIX" ]; then
 fi
 
 install -vm 644 build/bashrc  $PREFIX/etc/profile.d/infra.bash
+install -vm 644 build/profile  $PREFIX/etc/profile.d/infra.sh
 install -vm 644 build/inputrc $PREFIX/etc/inputrc
 install -vm 644 build/nanorc  $PREFIX/etc/nanorc
-install -vm 644 build/tmux.conf $PREFIX/etc/tmux.conf
 install -vm 644 build/tmux.conf $PREFIX/etc/tmux.conf
 
 mkdir  -pvm 755 $PREFIX/etc/nanorc.d/
@@ -23,7 +27,9 @@ install -vm 644 git/gitignore $PREFIX/etc/gitignore
 mkdir  -pvm 755 $BINPREFIX/bin/
 install -vm 755 tmux/tsend $BINPREFIX/bin/
 install -vm 755 tmux/tmux-popup-session $BINPREFIX/bin/
-install -vm 755 fzf/* $BINPREFIX/bin/
+if [ -f "/usr/bin/fzf" ]; then
+	install -vm 755 fzf/* $BINPREFIX/bin/
+fi
 install -vm 755 misc/* $BINPREFIX/bin/
 install -vm 755 nano/syntaxlist $BINPREFIX/bin/
 install -vm 644 htoprc $PREFIX/etc/htoprc
